% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_bantrafikenT04022.R
\name{fetch_bantrafikenT04022}
\alias{fetch_bantrafikenT04022}
\title{fetch_bantrafikenT04022}
\usage{
fetch_bantrafikenT04022(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_bantrafikenT04022
}
