% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_bussarT10011.R
\name{fetch_bussarT10011}
\alias{fetch_bussarT10011}
\title{fetch_bussarT10011}
\usage{
fetch_bussarT10011(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_bussarT10011
}
