% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_korkortT10012.R
\name{fetch_korkortT10012}
\alias{fetch_korkortT10012}
\title{fetch_korkortT10012}
\usage{
fetch_korkortT10012(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_korkortT10012
}
