% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lastbilar.R
\name{fetch_lastbilar}
\alias{fetch_lastbilar}
\title{fetch_lastbilar}
\usage{
fetch_lastbilar(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_lastbilar
}
