% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lastbilarT10033.R
\name{fetch_lastbilarT10033}
\alias{fetch_lastbilarT10033}
\title{fetch_lastbilarT10033}
\usage{
fetch_lastbilarT10033(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_lastbilarT10033
}
