% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lastbilarT10092.R
\name{fetch_lastbilarT10092}
\alias{fetch_lastbilarT10092}
\title{fetch_lastbilarT10092}
\usage{
fetch_lastbilarT10092(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_lastbilarT10092
}
