% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lastbilstrafik_kvartalT10062.R
\name{fetch_lastbilstrafik_kvartalT10062}
\alias{fetch_lastbilstrafik_kvartalT10062}
\title{fetch_lastbilstrafik_kvartalT10062}
\usage{
fetch_lastbilstrafik_kvartalT10062(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_lastbilstrafik_kvartalT10062
}
