% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mopedklass1T10015.R
\name{fetch_mopedklass1T10015}
\alias{fetch_mopedklass1T10015}
\title{fetch_mopedklass1T10015}
\usage{
fetch_mopedklass1T10015(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_mopedklass1T10015
}
