% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_motorcyklarT10014.R
\name{fetch_motorcyklarT10014}
\alias{fetch_motorcyklarT10014}
\title{fetch_motorcyklarT10014}
\usage{
fetch_motorcyklarT10014(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_motorcyklarT10014
}
