% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_personbilarT10016.R
\name{fetch_personbilarT10016}
\alias{fetch_personbilarT10016}
\title{fetch_personbilarT10016}
\usage{
fetch_personbilarT10016(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_personbilarT10016
}
