% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_personbilarT10094.R
\name{fetch_personbilarT10094}
\alias{fetch_personbilarT10094}
\title{fetch_personbilarT10094}
\usage{
fetch_personbilarT10094(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_personbilarT10094
}
