% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_postverksamhetT0701.R
\name{fetch_postverksamhetT0701}
\alias{fetch_postverksamhetT0701}
\title{fetch_postverksamhetT0701}
\usage{
fetch_postverksamhetT0701(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_postverksamhetT0701
}
