% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sjofartsforetagEuT08092.R
\name{fetch_sjofartsforetagEuT08092}
\alias{fetch_sjofartsforetagEuT08092}
\title{fetch_sjofartsforetagEuT08092}
\usage{
fetch_sjofartsforetagEuT08092(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_sjofartsforetagEuT08092
}
