% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sjofartsforetagT08091.R
\name{fetch_sjofartsforetagT08091}
\alias{fetch_sjofartsforetagT08091}
\title{fetch_sjofartsforetagT08091}
\usage{
fetch_sjofartsforetagT08091(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_sjofartsforetagT08091
}
