% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sjotrafikT0802.R
\name{fetch_sjotrafikT0802}
\alias{fetch_sjotrafikT0802}
\title{fetch_sjotrafikT0802}
\usage{
fetch_sjotrafikT0802(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_sjotrafikT0802
}
