% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_trafikarbeteT0401.R
\name{fetch_trafikarbeteT0401}
\alias{fetch_trafikarbeteT0401}
\title{fetch_trafikarbeteT0401}
\usage{
fetch_trafikarbeteT0401(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_trafikarbeteT0401
}
