% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_varufloden.R
\name{fetch_varufloden}
\alias{fetch_varufloden}
\title{fetch_varufloden}
\usage{
fetch_varufloden(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_varufloden
}
