% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_varuflodenT1102.R
\name{fetch_varuflodenT1102}
\alias{fetch_varuflodenT1102}
\title{fetch_varuflodenT1102}
\usage{
fetch_varuflodenT1102(URL, DATA_SELECTION)
}
\arguments{
\item{URL}{url to API}

\item{DATA_SELECTION}{number from 1 to N}
}
\description{
fetch_varuflodenT1102
}
