% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_lastbilstrafik_arT10061.R
\name{ow_lastbilstrafik_arT10061}
\alias{ow_lastbilstrafik_arT10061}
\title{ow_lastbilstrafik_arT10061}
\usage{
ow_lastbilstrafik_arT10061(URL)
}
\arguments{
\item{URL}{url to API}
}
\description{
ow_lastbilstrafik_arT10061
}
