% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_lastbilstrafik_kvartalT10062.R
\name{ow_lastbilstrafik_kvartalT10062}
\alias{ow_lastbilstrafik_kvartalT10062}
\title{ow_lastbilstrafik_kvartalT10062}
\usage{
ow_lastbilstrafik_kvartalT10062(URL)
}
\arguments{
\item{URL}{url to API}
}
\description{
ow_lastbilstrafik_kvartalT10062
}
