% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_plot_filter.R
\name{ow_plot_filter}
\alias{ow_plot_filter}
\title{ow_plot_filter}
\usage{
ow_plot_filter(df, VAR)
}
\arguments{
\item{df}{a df generated by ow_* function}

\item{VAR}{a Variable to filter for}
}
\description{
ow_plot_filter
}
