% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_postverksamhetT0701.R
\name{ow_postverksamhetT0701}
\alias{ow_postverksamhetT0701}
\title{ow_postverksamhetT0701}
\usage{
ow_postverksamhetT0701(URL)
}
\arguments{
\item{URL}{url to API}
}
\description{
ow_postverksamhetT0701
}
