% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_sjofartsforetagEuT08092.R
\name{ow_sjofartsforetagEuT08092}
\alias{ow_sjofartsforetagEuT08092}
\title{ow_sjofartsforetagEuT08092}
\usage{
ow_sjofartsforetagEuT08092(URL)
}
\arguments{
\item{URL}{url to API}
}
\description{
ow_sjofartsforetagEuT08092
}
