% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_sjofartsforetagT08091.R
\name{ow_sjofartsforetagT08091}
\alias{ow_sjofartsforetagT08091}
\title{ow_sjofartsforetagT08091}
\usage{
ow_sjofartsforetagT08091(URL)
}
\arguments{
\item{URL}{url to API}
}
\description{
ow_sjofartsforetagT08091
}
